#33 Moved Serializations namespace to NW.Shared.Serialization.
