+++ tests/NW.NGramTextClassification.UnitTests/NW.NGramTextClassification.UnitTests.csproj
@@ -17,6 +17,7 @@
   <ItemGroup>
     <PackageReference Include="nunit" Version="4.0.1" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.5.0" />
+    <PackageReference Include="NW.Shared.Serialization" Version="1.0.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\NW.NGramTextClassification\NW.NGramTextClassification.csproj" />
