+++ tests/NW.NGramTextClassification.UnitTests/Serialization/SerializerTests.cs
@@ -2,12 +2,12 @@
 using System.Collections.Generic;
 using NW.NGramTextClassification.LabeledExamples;
 using NW.NGramTextClassification.NGramTokenization;
-using NW.NGramTextClassification.Serializations;
-using NW.NGramTextClassification.TextSnippets;
-using NUnit.Framework;
 using NW.NGramTextClassification.TextClassifications;
+using NW.NGramTextClassification.TextSnippets;
+using NW.Shared.Serialization;
+using NUnit.Framework;
 
-namespace NW.NGramTextClassification.UnitTests.Serializations
+namespace NW.NGramTextClassification.UnitTests.Serialization
 {
     [TestFixture]
     public class SerializerTests
@@ -248,5 +248,5 @@ namespace NW.NGramTextClassification.UnitTests.Serializations
 
 /*
     Author: numbworks@gmail.com
-    Last Update: 31.01.2024
+    Last Update: 14.02.2024
 */
