+++ tests/NW.NGramTextClassification.UnitTests/TextClassifierTests.cs
@@ -10,16 +10,15 @@ using NW.NGramTextClassification.Bags;
 using NW.NGramTextClassification.LabeledExamples;
 using NW.NGramTextClassification.NGrams;
 using NW.NGramTextClassification.NGramTokenization;
-using NW.NGramTextClassification.Serializations;
 using NW.NGramTextClassification.Similarity;
 using NW.NGramTextClassification.TextClassifications;
 using NW.NGramTextClassification.TextSnippets;
 using NW.NGramTextClassification.UnitTests.Utilities;
 using NW.Shared.Files;
 using NW.Shared.Files.Validation;
+using NW.Shared.Serialization;
 using NW.Shared.Validation;
 using NUnit.Framework;
-using static System.Net.WebRequestMethods;
 
 namespace NW.NGramTextClassification.UnitTests
 {
@@ -1908,5 +1907,5 @@ namespace NW.NGramTextClassification.UnitTests
 
 /*
     Author: numbworks@gmail.com
-    Last Update: 12.02.2024
+    Last Update: 14.02.2024
 */
