+++ tests/NW.NGramTextClassificationClient.UnitTests/NW.NGramTextClassificationClient.UnitTests.csproj
@@ -17,6 +17,7 @@
   <ItemGroup>
     <PackageReference Include="NUnit" Version="4.0.1" />
     <PackageReference Include="NUnit3TestAdapter" Version="4.5.0" />
+    <PackageReference Include="NW.Shared.Serialization" Version="1.0.0" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\..\numbworks_logo_128x128px.png">
